+++ packages/node-firebird-driver/src/lib/impl/fb-util.ts
@@ -6,6 +6,7 @@ import * as stringDecoder from 'string_decoder';
 import { AbstractAttachment } from './attachment';
 import { AbstractTransaction } from './transaction';
 import { decodeDate, decodeTime, encodeDate, encodeTime } from './date-time';
+import { tzIdToString, tzStringToId } from './time-zones';
 
 import {
 	Attachment,
@@ -15,7 +16,8 @@ import {
 	CreateDatabaseOptions,
 	Transaction,
 	TransactionIsolation,
-	TransactionOptions
+	TransactionOptions,
+	ZonedDate
 } from '..';
 
 
@@ -35,6 +37,10 @@ export namespace sqlTypes {
 	export const SQL_TYPE_TIME = 560;
 	export const SQL_TYPE_DATE = 570;
 	export const SQL_INT64 = 580;
+	export const SQL_TIMESTAMP_TZ_EX = 32748;
+	export const SQL_TIME_TZ_EX = 32750;
+	export const SQL_TIMESTAMP_TZ = 32754;
+	export const SQL_TIME_TZ = 32756;
 	export const SQL_INT128 = 32752;
 	export const SQL_DEC16 = 32760;
 	export const SQL_DEC34 = 32762;
@@ -269,6 +275,20 @@ export function createDataReader(descriptors: Descriptor[]): DataReader {
 						decodedTime.hours, decodedTime.minutes, decodedTime.seconds, decodedTime.fractions / 10);
 				}
 
+				case sqlTypes.SQL_TIME_TZ:
+				case sqlTypes.SQL_TIME_TZ_EX: {
+					const now = new Date();
+					const decodedTime = decodeTime(dataView.getUint32(descriptor.offset, littleEndian));
+					const date = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth(), now.getUTCDate(),
+						decodedTime.hours, decodedTime.minutes, decodedTime.seconds, decodedTime.fractions / 10));
+					const timeZone = tzIdToString(dataView.getUint16(descriptor.offset + 4, littleEndian));
+
+					return {
+						date,
+						timeZone
+					} as ZonedDate;
+				}
+
 				case sqlTypes.SQL_TYPE_DATE: {
 					const decodedDate = decodeDate(dataView.getInt32(descriptor.offset, littleEndian));
 
@@ -295,6 +315,29 @@ export function createDataReader(descriptors: Descriptor[]): DataReader {
 						date.setFullYear(decodedDate.year);
 						return date;
 					}
+				}
+
+				case sqlTypes.SQL_TIMESTAMP_TZ:
+				case sqlTypes.SQL_TIMESTAMP_TZ_EX: {
+					const decodedDate = decodeDate(dataView.getInt32(descriptor.offset, littleEndian));
+					const decodedTime = decodeTime(dataView.getUint32(descriptor.offset + 4, littleEndian));
+					const timeZone = tzIdToString(dataView.getUint16(descriptor.offset + 8, littleEndian));
+					let date: Date;
+
+					if (decodedDate.year >= 100) {
+						date = new Date(Date.UTC(decodedDate.year, decodedDate.month - 1, decodedDate.day,
+							decodedTime.hours, decodedTime.minutes, decodedTime.seconds, decodedTime.fractions / 10));
+					}
+					else {
+						date = new Date(Date.UTC(2000, decodedDate.month - 1, decodedDate.day,
+							decodedTime.hours, decodedTime.minutes, decodedTime.seconds, decodedTime.fractions / 10));
+						date.setUTCFullYear(decodedDate.year);
+					}
+
+					return {
+						date,
+						timeZone
+					} as ZonedDate;
 				}
 
 				case sqlTypes.SQL_BOOLEAN:
@@ -389,6 +432,17 @@ export function createDataWriter(descriptors: Descriptor[]): DataWriter {
 					break;
 				}
 
+				case sqlTypes.SQL_TIME_TZ:
+				case sqlTypes.SQL_TIME_TZ_EX: {
+					const zonedDate = value as ZonedDate;
+					dataView.setUint32(descriptor.offset,
+						encodeTime(zonedDate.date.getUTCHours(), zonedDate.date.getUTCMinutes(), zonedDate.date.getUTCSeconds(),
+							zonedDate.date.getUTCMilliseconds() * 10),
+						littleEndian);
+					dataView.setUint16(descriptor.offset + 4, tzStringToId(zonedDate.timeZone), littleEndian);
+					break;
+				}
+
 				case sqlTypes.SQL_TYPE_DATE: {
 					const date = value as Date;
 					dataView.setInt32(descriptor.offset,
@@ -405,6 +459,20 @@ export function createDataWriter(descriptors: Descriptor[]): DataWriter {
 					dataView.setUint32(descriptor.offset + 4,
 						encodeTime(date.getHours(), date.getMinutes(), date.getSeconds(), date.getMilliseconds() * 10),
 						littleEndian);
+					break;
+				}
+
+				case sqlTypes.SQL_TIMESTAMP_TZ:
+				case sqlTypes.SQL_TIMESTAMP_TZ_EX: {
+					const zonedDate = value as ZonedDate;
+					dataView.setInt32(descriptor.offset,
+						encodeDate(zonedDate.date.getUTCFullYear(), zonedDate.date.getUTCMonth() + 1, zonedDate.date.getUTCDate()),
+						littleEndian);
+					dataView.setUint32(descriptor.offset + 4,
+						encodeTime(zonedDate.date.getUTCHours(), zonedDate.date.getUTCMinutes(), zonedDate.date.getUTCSeconds(),
+							zonedDate.date.getUTCMilliseconds() * 10),
+						littleEndian);
+					dataView.setUint16(descriptor.offset + 8, tzStringToId(zonedDate.timeZone), littleEndian);
 					break;
 				}